Remove generic_scheduler.go and generic_scheduler_test.go
